+++ Kernel/Modules/AgentSplitSelection.pm
@@ -116,11 +116,15 @@ sub Run {
     # Split Selection Dialog
     # ---------------------------------------------------------- #
 
-    # build split selection the modal dialog
-    my %SplitSelectionContent = (
-        PhoneTicket => Translatable('Phone ticket'),
-        EmailTicket => Translatable('Email ticket'),
-    );
+    # Build split selection based on availability of corresponding registered modules.
+    # See bug#13690 (https://bugs.otrs.org/show_bug.cgi?id=13690) for more information.
+    my %SplitSelectionContent;
+    if ( $ConfigObject->Get('Frontend::Module')->{AgentTicketPhone} ) {
+        $SplitSelectionContent{PhoneTicket} = Translatable('Phone ticket');
+    }
+    if ( $ConfigObject->Get('Frontend::Module')->{AgentTicketEmail} ) {
+        $SplitSelectionContent{EmailTicket} = Translatable('Email ticket');
+    }
 
     # only display process tickets if active processes are available
     my $Processes = $ConfigObject->Get('Process');